Removed duplicate test cases
