+++ src/interactions/test/services/test_comment_reactions.py
@@ -2,7 +2,6 @@ import pytest
 
 from interactions.models import CommentReaction, ReactionType
 from interactions.services.comment_reactions import (
-    get_active_reactions,
     react_to_comment,
     get_comment_reaction_count,
     get_comment_reaction_counts,
@@ -12,26 +11,6 @@ from interactions.services.comment_reactions import (
 from django.test import override_settings
 
 ALL_REACTION_TYPES = ReactionType.values
-
-
-@override_settings(INACTIVE_REACTION_TYPES=[])
-def test_get_active_reactions_default():
-    response = get_active_reactions()
-    assert ReactionType.CELEBRATE in response
-    assert ReactionType.LIKE in response
-    assert ReactionType.LOVE in response
-    assert ReactionType.DISLIKE in response
-    assert ReactionType.UNHAPPY in response
-
-
-@override_settings(INACTIVE_REACTION_TYPES=["unhappy", "like"])
-def test_get_active_reactions_settings():
-    response = get_active_reactions()
-    assert ReactionType.CELEBRATE in response
-    assert ReactionType.LIKE not in response
-    assert ReactionType.LOVE in response
-    assert ReactionType.DISLIKE in response
-    assert ReactionType.UNHAPPY not in response
 
 
 @pytest.mark.django_db
